requirements.txt
numpy
faiss-cpu
fastapi
pydantic
uvicorn